+++ Mozemoz/View Controllers/SearchVC/CollectionViewCell.swift
