+++ Mozemoz/View Controllers/SearchVC/SearchPageCollectionViewCell.swift
